requirements.txt
appdirs==1.4.3
backports.ssl-match-hostname==3.5.0.1
certifi==2017.4.17
httplib2==0.10.3
ipaddress==1.0.18
kafka-python==1.3.3
kubernetes==2.0.0a1
oauth2client==4.0.0
packaging==16.8
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
rsa==3.4.2
six==1.10.0
urllib3==1.20
websocket-client==0.40.0
